Merge develop into master. DAT-6437

+++ cloud-builder/docker-entrypoint.sh
@@ -104,6 +104,12 @@ echo "Deploying securely-waf to ${PROJECT_ID}..."
 docker pull "${SECURELY_WAF_IMAGE_SOURCE}"
 if [ -n "${ADDITIONAL_CA_CERT}" ]
 then
+    # shellcheck disable=SC2001
+    ADDITIONAL_CA_CERT=$(echo "${ADDITIONAL_CA_CERT}" | sed -e "s/^[[:space:]]*//")
+    if [ ! -f "${ADDITIONAL_CA_CERT}" ]; then
+        echo "File '${ADDITIONAL_CA_CERT}' cannot be found."
+        exit 1
+    fi
     mkdir -p waf_with_ca_cert
     cp "${ADDITIONAL_CA_CERT}" waf_with_ca_cert/
     pushd waf_with_ca_cert
